feat(cdk/dialog): move experimental CDK dialog into stable
Moves the `cdk-experimental/dialog` package into the CDK.
