+++ src/cdk/dialog/dialog-config.ts
