+++ src/cdk/dialog/dialog-module.ts
