+++ src/dev-app/dev-app/dev-app-layout.ts
@@ -36,10 +36,10 @@ export class DevAppLayout {
     {name: 'Button Toggle', route: '/button-toggle'},
     {name: 'Button', route: '/button'},
     {name: 'Card', route: '/card'},
-    {name: 'Cdk Dialog', route: '/cdk-dialog'},
-    {name: 'Cdk Experimental Combobox', route: '/cdk-experimental-combobox'},
-    {name: 'Cdk Experimental Listbox', route: '/cdk-experimental-listbox'},
-    {name: 'Cdk Experimental Menu', route: '/cdk-experimental-menu'},
+    {name: 'CDK Dialog', route: '/cdk-dialog'},
+    {name: 'CDK Experimental Combobox', route: '/cdk-experimental-combobox'},
+    {name: 'CDK Experimental Listbox', route: '/cdk-experimental-listbox'},
+    {name: 'CDK Experimental Menu', route: '/cdk-experimental-menu'},
     {name: 'Checkbox', route: '/checkbox'},
     {name: 'Chips', route: '/chips'},
     {name: 'Clipboard', route: '/clipboard'},
